### 2023-07-27 #5699 @editorialbot
[PRE REVIEW]: SPICY: A python toolbox for meshless assimilation from image velocimetry using radial basis functions
**Submitting author:** @mendezVKI (<a href="http://orcid.org/0000-0002-1115-2187">Miguel Alfonso Mendez</a>)
**Repository:** https://github.com/mendezVKI/SPICY_VKI
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
