### 2023-08-01 dc17651
Update paper.md
Minor English writing edits
