### 2023-08-01 #5699 @mendezVKI
Hello Philip,

Thank you for handling my submission. I am completely new to JOSS and its review process, so I might mess up in various ways.

I agree with all your English edits, of course. As for the reviewers, using the review checklist with (1) Python programming language, (2) Physics and Engineering as areas of expertise and (3) machine learning + fluid dynamics I get :

[Sven Berger](https://reviewers.joss.theoj.org/reviewers/2208)
[@akaptano](https://reviewers.joss.theoj.org/reviewers/1847)

Miguel
