### 2023-08-08 #5699 @philipcardiff
Hi @MatthewFlamm, would you be interested in reviewing this submission?

The submission under consideration is SPICY: A python toolbox for meshless assimilation from image velocimetry using radial basis functions (https://github.com/mendezVKI/SPICY_VKI).
