### 2023-08-09 #5699 @philipcardiff
Thanks, @MatthewFlamm.

Typically, you would post your initial comments within a few weeks, but longer is fine, too, if you are confident you can allocate the time.

For convenience, you can have a look at our [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) and [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html).
