### 2023-08-09 #5749 @editorialbot
[REVIEW]: SPICY: A python toolbox for meshless assimilation from image velocimetry using radial basis functions
**Submitting author:** @mendezVKI (<a href="http://orcid.org/0000-0002-1115-2187">Miguel Alfonso Mendez</a>)
**Repository:** https://github.com/mendezVKI/SPICY_VKI
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.5
