### 2023-08-09 mendezVKI/SPICY_VKI#2 @MatthewFlamm
Clean up unneeded files/folder; use appropriate .gitignore
In starting https://github.com/openjournals/joss-reviews/issues/5749, I noticed that there are several unneeded files/folders, probably caused by not using an appopriate `.gitignore` file for python repositories.

Not a requirement, but a recommendation would be to base your .gitignore file from a template available, for example: https://github.com/github/gitignore/blob/main/Python.gitignore.  You will want to add any specific file types for your particular workflow as well, e.g. `*.bak`.

Also not a requirement, but a recommendation: possible unnecessary files/folders that could be cleaned up if not needed are listed below. This is based on a scan and I did not assess whether these files truly are unneccessary.  I also did not do a thorough scan of the code yet.  Keeping the repo tidy makes it easier for contributors to understand the codebase and contribute.

```
.gitignore.bak
Tutorials/.ipynb_checkpoints
Tutorials/__pycache__
paper/paper.bib.sav.tmp
spicy_vki/build/lib/spicy  <- looks like it is mostly a copy of spicy_vki/spicy, which one is the one meant to be kept?
SPICY_Sphinx  <- looks to be similar to files in spicy_vki/docs
spicy_vki/spicy_vki.egg-info
spicy_vki/spicy/__pycache__
spicy_vki/spicy/spicy_class.py.bak
```
