### 2023-08-14 94be574
Added rtd.yaml file
Required file for building starting from 09/2023
