+++ spicy_vki/.readthedocs.yaml
@@ -31,4 +31,4 @@ sphinx:
 # See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
 python:
    install:
-   - requirements: spicy_vki.egg-info/requires.txt
+   - requirements: spicy_vki/spicy_vki.egg-info/requires.txt
