### 2023-08-14 f4a2a33
Update setup.py
Reading the requirements from the txt file
