### 2023-08-30 65fdc93
Fix installation instructions
