+++ spicy_vki/docs/source/installation.rst
@@ -4,4 +4,4 @@ Installation
 
 You can access the latest update of the spicy python package on PyPI. At the command line::
 
-    $ pip install spicy-vki=1.0.0
+    $ pip install spicy-vki==1.0.0
