### 2023-08-30 mendezVKI/SPICY_VKI#3 @MatthewFlamm
Are all dependencies required to run the software?
A lot of packages are installed when doing `pip install spicy-vki` that do not look neccessary for an end user, for example `numpydoc`, `sphinx`, `sphinxcontrib-serializinghtml`.  Consider removing `numpydoc` and any other non-needed packages from:

https://github.com/mendezVKI/SPICY_VKI/blob/f4a2a3376eeeef1e32471061c75b378d5ae4e131/spicy_vki/setup.py#L16-L24
