### 2023-08-30 mendezVKI/SPICY_VKI#3 @MatthewFlamm
ipython, numpydoc, and matplotlib all look to me only to be related to documentation building, but I haven't tested thoroughly.
