### 2023-08-30 #5749 @MatthewFlamm
Installation requires:
https://github.com/mendezVKI/SPICY_VKI/pull/4

The dependencies of the package are also a little messy, and this results in unneeded packages being installed, not as critical:
https://github.com/mendezVKI/SPICY_VKI/issues/3
