### 2023-08-30 mendezVKI/SPICY_VKI#3 @MatthewFlamm
Some of these dependencies are not be required for the package itself, but are used in the tutorials or documentation building.  In that case, there should be separate instructions for installing the packages or maybe a `requirements.txt` file that is available for installing the packages for the tutorials.  Another option would be to use optional dependencies like `pip install spciy-vki[tutorial]`.
