### 2023-08-31 0cf5601
Update paper.bib
Fix for reference
