+++ paper/paper.bib
@@ -1,7 +1,7 @@
 # References
-@InProceedings{Andrea,
+@InProceedings{Sciacchitano,
   author    = {Sciacchitano, A., Leclaire, B., Schr\"{o}der, A.},
-  booktitle = {20th International Symposium on the Application of Laser and Imaging Techniques to Fluid MEchanics},
+  booktitle = {20th International Symposium on the Application of Laser and Imaging Techniques to Fluid Mechanics},
   title     = {Main results of the first Data Assimilation Challenge},
   year      = {2022},
   address   = {Lisbon, Portugal},
