### 2023-08-31 mendezVKI/SPICY_VKI#8 @MatthewFlamm
Guidelines for users/contributors
Add guidelines for third parties somewhere, for example README, CONTRIBUTING.MD, or the documentation.  For reference the guidelines from https://github.com/openjournals/joss-reviews/issues/5749 are:

> Community guidelines
> There should be clear guidelines for third-parties wishing to:
>
> Contribute to the software
> Report issues or problems with the software
> Seek support

There are already some guidelines around running the content in the Tutorials, and issues related to that are already covered by #3 and #5 .  When those are addressed, adding in detail to the existing content explaining how to setup the environment for and run the Tutorials would be also be useful for users.
