### 2023-09-15 #5749 @MatthewFlamm
I have gone through the checklist and consider my review complete, pending needed changes to the paper and repository.  High level is that I believe this submission solves a useful scientific need.  The tutorials are of generally good quality, although I was not able to run the 3D case on my machine due to memory constraint.  In the future, I would suggest to the authors to think about how the package could become easier to use for users as the workflow seems to have many common elements in each case.  Today, the code does provide useful tools for researchers in this area. However, the repository is currently messily organized (see https://github.com/mendezVKI/SPICY_VKI/issues/2) and missing several key items for acceptance today.  See below on specifics.

Other than what I have already commented above, these are my comments relating to the last items:
Major comments:
- First reference is not handled correctly.  There is a PR that hopefully will address this, already linked above.
- Statement of need: is lacking a clear description of who the software is intended to be used by.
- State of the field: there is some discussion of recent applications, but PIV techniques have been used for many years.  The solution that this paper provides is at least adjacent to PIV techniques.  All the references are to scholarly publications, but I think the intent is to also discuss available software packages for these techniques.  There are several PIV software packages that are openly available, and I'm not sure if there are similar packages available more similar to SPCIY.  A discussion of the larger landscape including the available software packages would strengthen this paper greatly.

Minor comments:
- Automated tests: I have checked this box as there are tutorials which definitely increase the confidence in the functionality and accuracy of the code.  Within there are some hints to the user on how to verify the functionality, although some are qualitative.  A more dedicated test procedure would increase confidence in the package especially against regressions.
