### 2023-09-18 mendezVKI/SPICY_VKI#2 @mendezVKI
Solved!
