+++ spicy_vki/docs/source/installation.rst
@@ -5,3 +5,9 @@ Installation
 You can access the latest update of the spicy python package on PyPI. At the command line::
 
     $ pip install spicy-vki=1.0.0
+
+
+If you want to download also the folder containing the tutorials and the packages required
+to run them, the install also the optional dependencies as follows:
+
+    $ pip install spicy-vki=1.0.0[tutorials]
