+++ spicy_vki/spicy_vki.egg-info/requires.txt
@@ -1,7 +1,5 @@
 numpy>=1.20
-matplotlib>=3.3.0
 scikit-learn>=1.0
 ipython>=7.16.1
 scipy>=1.5
 shapely>=1.7.0
-numpydoc>=1.1.0
