+++ README.md
@@ -8,13 +8,13 @@ The framework works for structured data (as produced by cross-correlation-based 
 While the main scope is the assimilation of velocity fields, SPICY can also be used for the regression of other fields (e.g., temperature fields).
 The theoretical foundation of the constrained RBF approach is described in - P. Sperotto, S. Pieraccini, M.A. Mendez, A Meshless Method to Compute Pressure Fields from Image Velocimetry, Measurement Science and Technology 33(9), May 2022. (pre-print at https://arxiv.org/abs/2112.12752).
 
-The GitHub folder contains four exercises in Python. These include regression of synthetic velocity fields as well as the solution of Poisson problems.
+The GitHub folder contains four tutorials. These include regression of synthetic velocity fields as well as the solution of Poisson problems.
 For each tutorial, there is a video on our YouTube channel: https://www.youtube.com/@spicyVKI
 
 The documentatation can be found here: https://spicy-vki.readthedocs.io/en/latest/index.html
 
 
-The list of proposed exercises is following:
+The list of proposed tutorials is the following:
 
 1 - Solution of a Laplace problem on the unit square.
 
