### 2023-09-18 3d75acd
update readme + add contribuing
