+++ contribuing.md
@@ -1,0 +1,37 @@
+Contributing
+========
+
+Contribute to SPICY
+-----
+If you want to contribute to the project, don't hesitate to Pull Request. Please provide a broad description of the improvement you intend to implement and contact the mantainers.
+
+
+Report Issues
+------
+
+If you come across a bug, have a feature request, or notice any unexpected behavior, please open an issue on this repository. To do so, follow these steps:
+
+1. 	Click on the "Issues" tab at the top of this repository's page.
+2.  Click the "New Issue" button.
+3.  Provide a descriptive title for the issue.
+4.  Detail the problem or request in the issue description, including steps to reproduce it if applicable.
+5.  Use labels and assignees to categorize and direct the issue as needed.
+
+We will address them as soon as we can.
+
+Seek Support
+-----
+
+SPICY has some tutorials to get started with the package, available at: `SPICY_VKI/Tutorials`
+If you want to run those please ensure that you install the optional dependencies required via: 
+`pip install spicy_vki[tutorials]`
+
+If you have any additional question on how to use the package, please don't hesitate to contact us. However, we recommend first to check:
+
+- Documentation: Check the README and any other relevant documentation in this repository. Many common questions are answered there.
+
+- Search Existing Issues: The issue tracker may already contain discussions or solutions related to your problem. Be sure to search through existing issues to see if your question has been addressed.
+
+- GitHub Discussions: If you have general questions, ideas, or want to discuss topics related to this repository, feel free to use the GitHub Discussions feature. It's a great place for open conversations.
+
+- Contact the Maintainer: If you cannot find a solution through the above methods, you can contact the repository maintainer directly. Look for contact information in the repository's documentation.
