+++ spicy_vki/docs/source/installation.rst
@@ -6,7 +6,6 @@ You can access the latest update of the spicy python package on PyPI. At the com
 
     $ pip install spicy-vki=1.0.0
 
-
 If you want to download also the folder containing the tutorials and the packages required
 to run them, the install also the optional dependencies as follows:
 
