+++ NOTE_BUILD.txt
@@ -1,10 +1,0 @@
-
-
-To build the article:
-
-
-To build the documentation: 
-
-
-To build the installation package :
-
