### 2023-09-18 mendezVKI/SPICY_VKI#3 @mendezVKI
I removed "numpydoc>=1.1.0". Concerning the tutorials, we followed your suggestion and created an installation with optional dependencies. This was also updated in the documentation.
