### 2023-09-18 mendezVKI/SPICY_VKI#8 @mendezVKI
We added a contribuing.md file with instruction for contributing
