+++ paper/paper.bib
@@ -6,6 +6,8 @@
   year      = {2022},
   address   = {Lisbon, Portugal},
 }
+
+
 @Article{Sperotto2022,
   author    = {Pietro Sperotto and Sandra Pieraccini and Miguel A Mendez},
   journal   = {Measurement Science and Technology},
