+++ paper/paper.md
@@ -45,8 +45,7 @@ A set of video tutorials on how to use SPICY is provided.
 Data assimilation methods are becoming increasingly crucial in image velocimetry, thanks to high spatial and temporal resolutions available with modern interrogation processing methods. Assimilation techniques optimally combine measurements and first principle models (e.g. conservation laws) to maximize noise removal, achieve measurement super-resolution, and compute related quantities such as vorticity and pressure fields. Several methods have been proposed in the literature and assessed in the framework of the European project HOMER (Holistic Optical Metrology for Aero-Elastic Research), grant agreement number 769237. The most classic approaches for the velocity assimilation involve regression of the particle trajectories (as in TrackFit by [@Gesemann2016]), while the computation of derived quantities is usually carried out by first mapping the tracks onto Eulerian grids and then solving the relevant PDEs using standard CFD approaches (as in [@Schneiders2016a, @Agarwal2021] ).  
 
 Alternatives mesh-free methods are the second-generation Flowfit [@Gesemann2016], which combines the velocity regression and the pressure integration into a large nonlinear optimization problem, and methods based on physics-informed neural networks (PINNs) [@Rao2020,] which uses penalized artificial neural networks to solve for the velocity and the pressure fields. 
-Recently, in [@Sperotto2022], we proposed a meshless approach based on constrained Radial Basis Functions (RBFs) to solve both the velocity regression problem and the pressure computation. This approach is akin to the well-known Kansa method [@Fornberg2015] for the meshless integration of PDEs. The main novelty is that this formulation yields linear regression problems that can be easily constrained (rather than penalized) and solved using standard linear system solvers. All the codes developed have now been released in the open SPICY (Super Resolution and Pressure from Image Velocimetry) toolbox linked to this contribution. Documentation, installation, and tutorials are available in the provided repository and on a [Youtube channel](https://youtu.be/VYeiip_mEtg). 
-While several open-source codes are available for PTV[@Sperotto2022] and PIV [@Sperotto2022], to the author knowledge SPICY is the first opensource code for assimilation and pressure computation from image velocimetry.
+Recently, in [@Sperotto2022], we proposed a meshless approach based on constrained Radial Basis Functions (RBFs) to solve both the velocity regression problem and the pressure computation. This approach is akin to the well-known Kansa method [@Fornberg2015] for the meshless integration of PDEs. The main novelty is that this formulation yields linear regression problems that can be easily constrained (rather than penalized) and solved using standard linear system solvers. All the codes developed have now been released in the open SPICY (Super Resolution and Pressure from Image Velocimetry) toolbox linked to this contribution. Documentation, installation, and tutorials are available in the provided repository and on a [Youtube channel](https://youtu.be/VYeiip_mEtg). While several open-source codes are available for PTV[Heyman2019,Meller2016] and PIV [@Liberzon2020,Thielicke2021], to the author knowledge SPICY is the first opensource code for assimilation and pressure computation from image velocimetry.
 
 # Tutorials and ongoing works
 
