+++ paper/paper.bib
@@ -21,6 +21,60 @@
   publisher = {{IOP} Publishing},
 }
 
+
+
+
+@Article{Heyman2019,
+  author    = {Joris Heyman},
+  journal   = {Computers {\&}amp$\mathsemicolon$ Geosciences},
+  title     = {{TracTrac}: A fast multi-object tracking algorithm for motion estimation},
+  year      = {2019},
+  month     = {jul},
+  pages     = {11--18},
+  volume    = {128},
+  doi       = {10.1016/j.cageo.2019.03.007},
+  publisher = {Elsevier {BV}},
+}
+
+@Article{Meller2016,
+  author    = {Yosef Meller and Alex Liberzon},
+  journal   = {Journal of Open Research Software},
+  title     = {Particle Data Management Software for 3DParticle Tracking Velocimetry and Related Applications {\textendash} The Flowtracks Package},
+  year      = {2016},
+  month     = {jun},
+  number    = {1},
+  pages     = {23},
+  volume    = {4},
+  doi       = {10.5334/jors.101},
+  publisher = {Ubiquity Press, Ltd.},
+}
+
+@Misc{Liberzon2020,
+  author    = {Liberzon, Alex and {Roi Gurka} and Lepchev, Denis and {Hadar Ben-Gida}},
+  title     = {OpenPIV Spatial and Temporal Analysis Toolbox},
+  year      = {2020},
+  copyright = {MIT License},
+  doi       = {10.6084/M9.FIGSHARE.12330608},
+  keywords  = {Mechanical Engineering, FOS: Mechanical engineering, FOS: Mechanical engineering, 91504 Fluidisation and Fluid Mechanics, FOS: Other engineering and technologies, FOS: Other engineering and technologies, 40403 Geophysical Fluid Dynamics, FOS: Earth and related environmental sciences, FOS: Earth and related environmental sciences, 20303 Fluid Physics, FOS: Physical sciences, FOS: Physical sciences, 91501 Computational Fluid Dynamics, 90101 Aerodynamics (excl. Hypersonic Aerodynamics)},
+  publisher = {figshare},
+}
+
+@Article{Thielicke2021,
+  author    = {William Thielicke and Ren{\'{e}} Sonntag},
+  journal   = {Journal of Open Research Software},
+  title     = {Particle Image Velocimetry for {MATLAB}: Accuracy and enhanced algorithms in {PIVlab}},
+  year      = {2021},
+  month     = {may},
+  number    = {1},
+  pages     = {12},
+  volume    = {9},
+  doi       = {10.5334/jors.334},
+  publisher = {Ubiquity Press, Ltd.},
+}
+
+
+
+
 @InProceedings{Ratz2022a,
   author    = {Ratz, M., Sachs, S., K\"{o}nig, J., Mendez, M. A., Cierpka, C.},
   booktitle = {International Symposium on Applications of Laser and Imaging Techniques to Fluid Mechanics},
@@ -89,3 +143,4 @@
   doi       = {10.1007/s00348-021-03172-0},
   publisher = {Springer Science and Business Media {LLC}},
 }
+
