### 2023-09-18 #5749 @mendezVKI
Dears, thank you so much for your detailed review.

I have addressed all points in the issues. Concerning the ones open in this check list:

1. References: I modified a bit them, removing the one that was creating problem and added others concerning open-software for PIV/PTV. There are still a couple of references that do not have a DOI; this is because these are conf papers. I considered for a moment removing these, but I feel that there is very useful information there so I prefer to keep them.

2. A statement of need: I add a sentence mentioning open-source software and about positioning SPICY in the larger landscape. Although some authors (e.g. the cited Rao 2020) shared a github ref with their implementation of the assimilation using PINNs, we have not found any software package similar to SPICY. We state this clearly. Concerning the lack of "a clear description of who the software is intended to be used by", we added a clear statement in the readme but can't fit it in the 2 page format of the article without removing important references. The section reports the main context (data assimilation for velocimetry) that the SPICY package targets, so it is implicit that it targets anyone interested in this topic.

3. Automated tests: We will consider this for the next release; thank you very much for the suggestion. As beginners in software developments we limited ourselves to tutorials designed to test all the key functions. In our experience, it seems that users learn how to use the code quite quickly while having a good grasp of the underlying numerics from the videos we provided on youtube (https://www.youtube.com/@spicyVKI).

4.  Community guidelines. Indeed, these were missing. We added an .md file with instructions (as per solved issue)

5. Installation instructions and related documentation. Indeed, several dependencies were unclear. We fixed this as suggested in the issue. We now include an installation with optional dependencies for the tutorials.

The folder should now be clean. I twine upload everything in a few hours.

Once again, thank you so much for the very detailed review and suggestions :)
