### 2023-10-27 #5749 @philipcardiff
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1088/1361-6501/ac70a9 is OK
> - 10.1016/j.cageo.2019.03.007 is OK
> - 10.5334/jors.101 is OK
> - 10.6084/M9.FIGSHARE.12330608 is OK
> - 10.5334/jors.334 is OK
> - 10.1017/s0962492914000130 is OK
> - 10.1007/s00348-016-2133-9 is OK
> - 10.1007/s00348-021-03172-0 is OK
>
> MISSING DOIs
>
> - 10.18409/ispiv.v1i1.198 may be a valid DOI for title: Main results of the first Data Assimilation Challenge
> - 10.55037/lxlaser.20th.115 may be a valid DOI for title: A RANS approach to the Meshless Computation of Pressure Fields From Image Velocimetry
>
> INVALID DOIs
>
> - https://doi.org/10.1016/j.taml.2020.01.039 is INVALID because of 'https://doi.org/' prefix

Hi @mendezVKI, please fix these missing and invalid references. Thanks!
