### 2023-11-14 #5749 @mendezVKI
I apologize for the late response; I missed this update.

The reference "Main results of the first Data Assimilation Challenge" has been removed.  Did something go wrong with the resubmission ? I attach the latest pdf I find in my "Action-> Artifacts" session
[paper.pdf](https://github.com/openjournals/joss-reviews/files/13346898/paper.pdf)

The references that do not have a DOI are conference papers for which this info is not relevant. Nevertheless, these are still very valuable to this work and we wish to cite these (papers are available on arxiv).
