+++ paper/paper.bib
@@ -106,7 +106,7 @@
     year     = {2020},
     number   = {3},
     volume   = {10},
-    doi      = {https://doi.org/10.1016/j.taml.2020.01.039},
+    doi      = {10.1016/j.taml.2020.01.039},
     keywords = {Physics-informed neural networks (PINN), Deep learning, Fluid dynamics, Incompressible laminar flow},
 }
 
