### 2023-11-15 31c54b3
Fixed consistent notation of the distance norms
