+++ spicy_vki/spicy/spicy_class.py
@@ -114,8 +114,8 @@ class spicy:
         # Check the input is correct
         assert type(data) == list, 'Input data must be a list'
         assert type(grid_point) == list, 'Input grid_point must be a list'
-        assert type(basis) == str, 'Basis must be a string'
-        assert ST == None or type(ST) == list, 'ST must be a string or a list'
+        assert type(basis) == str, 'Input basis must be a string'
+        assert ST == None or type(ST) == list, 'Input ST must be a \'None\' or a list'
                         
         # Assign the basis
         if basis == 'gauss' or basis == 'c4':
@@ -1967,24 +1967,24 @@ def Phi_RBF_2D(X_G, Y_G, X_C, Y_C, c_k, basis):
     """
     # This is the contribution of the RBF part
     n_b = len(X_C); n_p = len(X_G)
-    Phi_RBF = np.zeros((n_p,n_b))
+    Phi_RBF = np.zeros((n_p, n_b))
     
     # What comes next depends on the type of chosen RBF
     if basis == 'gauss':
         # Iterate over all basis elements
         for r in range(n_b):
             # Compute the Gaussian
-            gaussian=np.exp(-c_k[r]**2*((X_C[r]-X_G)**2+(Y_C[r]-Y_G)**2))
+            gaussian = np.exp(-c_k[r]**2 * ((X_G-X_C[r])**2 + (Y_G-Y_C[r])**2))
             # Assemble into matrix
-            Phi_RBF[:,r]=gaussian
+            Phi_RBF[:,r] = gaussian
 
     elif basis == 'c4':
         # Iterate over all basis elements
         for r in range(n_b):
             # Get distance
-            d = np.sqrt((X_G - X_C[r])**2 + (Y_G - Y_C[r])**2)
+            d = np.sqrt((X_G-X_C[r])**2 + (Y_G-Y_C[r])**2)
             # Compute Phi
-            phi = (1 + d/c_k[r])**5 * (1 - d/c_k[r])**5
+            phi = (1+d/c_k[r])**5 * (1-d/c_k[r])**5
             # Compact support
             phi[np.abs(d) > c_k[r]] = 0
             # Assemble into matrix
@@ -2003,24 +2003,24 @@ def Phi_RBF_2D_x(X_G, Y_G, X_C, Y_C, c_k, basis):
     # number of bases (n_b) and points (n_p)
     n_b = len(X_C); n_p = len(X_G)
     # Initialize the matrix
-    Phi_RBF_x = np.zeros((n_p,n_b))
+    Phi_RBF_x = np.zeros((n_p, n_b))
     
     # What comes next depends on the type of chosen RBF
     if basis == 'gauss':
         # Iterate over all basis elements
         for r in range(n_b):
             # Compute the Gaussian
-            gaussian=np.exp(-c_k[r]**2*((X_C[r]-X_G)**2+(Y_C[r]-Y_G)**2))
+            gaussian = np.exp(-c_k[r]**2 * ((X_G-X_C[r])**2 + (Y_G-Y_C[r])**2))
             # Multiply with inner term and assemble into matrix
-            Phi_RBF_x[:,r]=2*c_k[r]**2*(X_C[r]-X_G)*gaussian
+            Phi_RBF_x[:,r] = - 2*c_k[r]**2 * (X_G-X_C[r])*gaussian
             
[…]
