### 2023-11-16 mendezVKI/SPICY_VKI#9 @MatthewFlamm
installation instructions in documentation typo
As in #4, the following line

https://github.com/mendezVKI/SPICY_VKI/blob/31c54b3005ff2adccca1e80aa643b8406fc6ac1d/spicy_vki/docs/source/installation.rst?plain=1#L7

Needs to be

     $ pip install spicy-vki==1.0.0

See https://packaging.python.org/en/latest/specifications/version-specifiers/#id4

There is a line below this that also needs to be updated.
