### 2023-11-16 mendezVKI/SPICY_VKI#9 @MatthewFlamm
Actually this version is already out of date with your changes.  I recommend just removing the version entirely. That is:

```
 $ pip install spicy-vki
```

just like you have in the readme.
