### 2023-11-16 #5749 @MatthewFlamm
I got busy in the last few weeks, but was able to get back to this.  I have checked all the boxes.

While there is no explicit statement of who the software is intended to be used by, I think the article and tutorials provide enough context, especially given the expanded discussion of other software packages.

I have added an issue https://github.com/mendezVKI/SPICY_VKI/issues/9, which highlights a typo for installation instructions. But given the substantial changes to the readme and clear installation instructions there, this doesn't hold up my approval on this point.
