### 2023-11-20 mendezVKI/SPICY_VKI#9 @mendezVKI
Solved !
