+++ spicy_vki/.readthedocs.yaml
@@ -30,8 +30,7 @@ sphinx:
 #Optional but recommended, declare the Python requirements required to build your documentation
 # See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
 python:
+   version: 3.8
    install:
-   - method: pip
      requirements: spicy_vki/spicy_vki.egg-info/requires.txt
-     extra_requirements:
-        - tutorials
+     extra_requirements: [tutorials]
