### 2023-11-21 eeec22b
Update .readthedocs.yaml
