+++ spicy_vki/.readthedocs.yaml
@@ -4,32 +4,14 @@
 # Required
 version: 2
 
-# Set the OS, Python version and other tools you might need
-build:
-  os: ubuntu-22.04
-  tools:
-    python: "3.12"
-    # You can also specify other tool versions:
-    # nodejs: "20"
-    # rust: "1.70"
-    # golang: "1.20"
-
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
   configuration: spicy_vki/docs/source/conf.py
-  # You can configure Sphinx to use a different builder, for instance use the dirhtml builder for simpler URLs
-  # builder: "dirhtml"
-  # Fail on all warnings to avoid broken references
-  # fail_on_warning: true
-
-# Optionally build your docs in additional formats such as PDF and ePub
-# formats:
-#    - pdf
-#    - epub
 
 #Optional but recommended, declare the Python requirements required to build your documentation
 # See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
 python:
+   version: 3.11
    install:
      - requirements: spicy_vki/spicy_vki.egg-info/requires.txt
        extra_requirements: [tutorials]
