+++ spicy_vki/.readthedocs.yaml
@@ -18,4 +18,5 @@ python:
        
 build:
   os: ubuntu-22.04
-  image: latest
+  tools:
+    python: "3.11"
