### 2023-11-21 a5137f5
Revert "Update .readthedocs.yaml"
This reverts commit da429b2440df84ad7f5cb31a04d237c27c57d082.
