+++ spicy_vki/.readthedocs.yaml
@@ -4,11 +4,6 @@
 # Required
 version: 2
 
-build:
-  os: ubuntu-22.04
-  tools:
-    python: "3.11"
-    
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
   configuration: spicy_vki/docs/source/conf.py
@@ -18,8 +13,9 @@ sphinx:
 python:
    install:
      - requirements: spicy_vki/spicy_vki.egg-info/requires.txt
-     - method: pip
-       path: .
-       extra_requirements:
-         - [tutorials]
+       extra_requirements: [tutorials]
        
+build:
+  os: ubuntu-22.04
+  tools:
+    python: "3.11"
