### 2023-11-21 5d98c73
Revert "Update .readthedocs.yaml"
This reverts commit e5de84adc8e11d0715e41a0ab342d972730d9ac4.
