+++ spicy_vki/.readthedocs.yaml
@@ -11,6 +11,7 @@ sphinx:
 #Optional but recommended, declare the Python requirements required to build your documentation
 # See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
 python:
+   version: 3.11
    install:
      - requirements: spicy_vki/spicy_vki.egg-info/requires.txt
        extra_requirements: [tutorials]
