### 2023-11-21 d92a504
Revert "Update .readthedocs.yaml"
This reverts commit fa48883b3927e2565aa46d7b82bb2fdee1a346cc.
