+++ spicy_vki/.readthedocs.yaml
@@ -18,5 +18,4 @@ python:
        
 build:
   os: ubuntu-22.04
-  tools:
-    python: "3.11"
+  image: latest
