### 2023-11-21 3fe0742
Revert "Update .readthedocs.yaml"
This reverts commit f5dd4dbfb398dca97b6fd652c3a73c4bb86a8c7e.
