### 2023-11-21 5e5548b
Revert "Update .readthedocs.yaml"
This reverts commit eeec22b4217ef573d0abb64934f88fa98cba656a.
