### 2023-11-21 e5af19a
Revert "Update .readthedocs.yaml"
This reverts commit dd8fd452f6a0539f21eb9cfbabc834f7833063bf.
