+++ spicy_vki/.readthedocs.yaml
@@ -7,8 +7,6 @@ version: 2
 # Set the OS, Python version and other tools you might need
 build:
   os: ubuntu-22.04
-  tools:
-    python: "3.12"
     # You can also specify other tool versions:
     # nodejs: "20"
     # rust: "1.70"
@@ -30,6 +28,7 @@ sphinx:
 #Optional but recommended, declare the Python requirements required to build your documentation
 # See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
 python:
+   version: 3.11
    install:
-     - requirements: spicy_vki/spicy_vki.egg-info/requires.txt
-       extra_requirements: [tutorials]
+     requirements: spicy_vki/spicy_vki.egg-info/requires.txt
+     extra_requirements: [tutorials]
