### 2023-11-21 3f2c8ae
Revert "Update .readthedocs.yaml"
This reverts commit 93246aff54dd4dd4bd09429de36d5a98e7a8aa47.
