### 2023-11-21 3309709
Revert "Update .readthedocs.yaml"
This reverts commit 2d2e971c006f76a35d3a08f5f252e2f3c82d34f6.
