### 2023-11-21 ac6b034
Revert "Update .readthedocs.yaml"
This reverts commit 370ac34808f3ee2e3d26942c46d7c23f7d9779dd.
