+++ spicy_vki/.readthedocs.yaml
@@ -31,7 +31,6 @@ sphinx:
 # See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
 python:
    install:
-   - method: pip
-     requirements: spicy_vki/spicy_vki.egg-info/requires.txt
+   - requirements: spicy_vki/spicy_vki.egg-info/requires.txt
      extra_requirements:
-        - tutorials
+     - tutorials
