### 2023-11-21 fedd314
Revert "Update .readthedocs.yaml"
This reverts commit 8f199c798e7816d24a91917b3d1c9d046bc5ada2.
