+++ spicy_vki/.readthedocs.yaml
@@ -8,7 +8,7 @@ version: 2
 build:
   os: ubuntu-22.04
   tools:
-    python: "3.11"
+    python: "3.8"
     # You can also specify other tool versions:
     # nodejs: "20"
     # rust: "1.70"
@@ -32,5 +32,3 @@ sphinx:
 python:
    install:
    - requirements: spicy_vki/spicy_vki.egg-info/requires.txt
-     extra_requirements:
-     - tutorials
