### 2023-11-21 1f04319
Revert "Update .readthedocs.yaml"
This reverts commit 51ee16874ed318b49b1e52d1e612fc58e32fe3c6.
