### 2023-11-21 b5c0568
Fixed the optional dependencies which broke the new readthedocs builds
