### 2023-11-22 7c187ff
Fixed the computation of the shape parameters.
The shape parameters were computed wrong such that the RBFs have a more narrow shape. The results do not change, if simply the clipping of the radii is doubled. This is fixed now.
