### 2023-11-22 7a11887
Updated spicy version to 1.0.2
