+++ spicy_vki/build/lib/spicy/spicy_class.py
@@ -49,10 +49,6 @@ class spicy:
     def __init__(self, data, grid_point, basis='gauss', ST=None):
         """
         Initialization of an instance of the spicy class. 
-        
-        :type model: str
-        :param model: 
-            This defines the model. Currently, SPICY supports 4 models:
                                           
         :type data: list of 1D numpy.ndarray
         :param data:
@@ -118,8 +114,8 @@ class spicy:
         # Check the input is correct
         assert type(data) == list, 'Input data must be a list'
         assert type(grid_point) == list, 'Input grid_point must be a list'
-        assert type(basis) == str, 'Basis must be a string'
-        assert ST == None or type(ST) == list, 'ST must be a string or a list'
+        assert type(basis) == str, 'Input basis must be a string'
+        assert ST == None or type(ST) == list, 'Input ST must be a \'None\' or a list'
                         
         # Assign the basis
         if basis == 'gauss' or basis == 'c4':
@@ -351,27 +347,27 @@ class spicy:
         # on the type of RBF but not whether the type is 2D or 3D.
         if self.basis =='gauss':
             # Set the max and min values of c_k
-            c_min = 1/(2*r_mM[1])*np.sqrt(np.log(2))
-            c_max = 1/(2*r_mM[0])*np.sqrt(np.log(2))
+            c_min = 1/(r_mM[1])*np.sqrt(np.log(2))
+            c_max = 1/(r_mM[0])*np.sqrt(np.log(2))
             # compute the c_k 
             c_k = np.sqrt(-np.log(eps_l))/sigma
             # crop to the minimum and maximum value
             c_k[c_k < c_min] = c_min
             c_k[c_k > c_max] = c_max
             # for plotting purposes, we store also the diameters
-            d_k = 1/c_k*np.sqrt(np.log(2))
+            d_k = 2/c_k*np.sqrt(np.log(2))
             
         elif self.basis == 'c4':
             # Set the max and min values of c_k
-            c_min = 2*r_mM[0] / np.sqrt(1 - 0.5**0.2)
-            c_max = 2*r_mM[1] / np.sqrt(1 - 0.5**0.2)
+            c_min = r_mM[0] / np.sqrt(1 - 0.5**0.2)
+            c_max = r_mM[1] / np.sqrt(1 - 0.5**0.2)
             # compute the c _k
             c_k = sigma / np.sqrt(1 - eps_l**0.2)
             # crop to the minimum and maximum value
             c_k[c_k < c_min] = c_min
             c_k[c_k > c_max] = c_max
             # for plotting purposes, we store also the diameters
-            d_k = c_k * np.sqrt(1 - 0.5**0.2)
+            d_k = 2*c_k * np.sqrt(1 - 0.5**0.2)
         self.c_k = c_k
         self.d_k = d_k
         
@@ -1091,7 +1087,7 @@ class spicy:
                                self.c_k, self.basis)
                     ))
                 # Assemble A and b_1, we also rescale b_1
[…]
