### 2023-11-22 a4c9dc6
Updated documentation to reflect the changes from version 1.0.0 to 1.0.2
