+++ spicy_vki/docs/build/html/_sources/version_history.rst.txt
@@ -1,6 +1,15 @@
 ===============
 Release History
 ===============
+2023-11-11 - `spicy_vki == 1.0.2`
+-----------------------------------
+- Bugfix of the shape parameters. The shape parameters were computed such that the RBFs were half as big as intended. This is now fixed. The tutorials are updated, but old tutorials and other code still yield the same result when the minimum and maximum diameter are given twice as large.
+- Fixed consistent notation in the computation of distances. All computations are now done with :math:`d = \sqrt{\boldsymbol{X} - \boldsymbol{X}_C^2}`
+
+
+2023-09-20 - `spicy_vki == 1.0.1`
+-----------------------------------
+- Bugfix where a 3D scalar regression threw an error
 
 2023-06-13 - `spicy_vki == 1.0.0`
 -----------------------------------
