+++ spicy_vki/spicy_vki.egg-info/requires.txt
@@ -4,3 +4,5 @@ ipython>=7.16.1
 scipy>=1.5
 shapely>=1.7.0
 matplotlib>=3.3.0
+sphinx_rtd_theme>=1.3.0
+numpydoc>=1.6.0
