### 2023-11-22 2715cf8
Minor  bug fix in 1.0.3 where readthedocs was not building correctly
