+++ spicy_vki/docs/source/conf.py
@@ -25,8 +25,8 @@ author = 'P. Sperotto, M. Ratz, M. A. Mendez'
 import spicy_class
 
 # The full version, including alpha/beta/rc tags
-version = '1.0.2'
-release = '1.0.2'
+version = '1.0.3'
+release = '1.0.3'
 
 # -- General configuration ---------------------------------------------------
 
