### 2023-11-22 285fe62
Removed spicy_vki.egg-info from the gitignore and patched it for 1.0.4
