+++ spicy_vki/docs/source/conf.py
@@ -12,7 +12,6 @@
 #
 import os
 import sys
-# sys.path.insert(0, os.path.abspath('.'))
 sys.path.append(os.path.abspath('..' + os.sep + '..' + os.sep + 'spicy'))
 
 
