+++ spicy_vki/spicy_vki.egg-info/SOURCES.txt
@@ -1,3 +1,4 @@
+README.md
 setup.py
 spicy/__init__.py
 spicy/spicy_class.py
