### 2023-12-28 #5749 @mendezVKI
Hello @philipcardiff ,

I hope you are having a wonderful Christmas break.

Do you have an estimation of the timing before final publication? I am asking because we would like to cite this article in several upcoming works that used the software.

All the best

Miguel
