### 2024-01-04 #5749 @nolankucd
> On 9 Aug 2023, at 14:49, The Open Journals editorial robot ***@***.***> wrote:
>
>
> Hello humans, I'm @editorialbot <https://github.com/editorialbot>, a robot that can help you with some common editorial tasks.
>
> For a list of things I can do to help you, just type:
>
> @editorialbot commands
> For example, to regenerate the paper pdf after making changes in the paper's md or bib files, type:
>
> @editorialbot generate pdf
