### 2024-01-08 #5749 @philipcardiff
> Hello @philipcardiff ,
>
> I hope you are having a wonderful Christmas break.
>
> Do you have an estimation of the timing before final publication? I am asking because we would like to cite this article in several upcoming works that used the software.
>
> All the best
>
> Miguel

Hi Moguel,

Apologies for the delay. Once @nolankucd confirms his review is complete, we can move to the final publication stages, which should be relatively quick.

Philip
