+++ paper/paper.bib
@@ -1,3 +1,4 @@
+% Encoding: UTF-8
 # References
 @InProceedings{Sciacchitano,
   author    = {Sciacchitano, A., Leclaire, B., Schr\"{o}der, A.},
@@ -55,7 +56,7 @@
   year      = {2020},
   copyright = {MIT License},
   doi       = {10.6084/M9.FIGSHARE.12330608},
-  keywords  = {Mechanical Engineering, FOS: Mechanical engineering, FOS: Mechanical engineering, 91504 Fluidisation and Fluid Mechanics, FOS: Other engineering and technologies, FOS: Other engineering and technologies, 40403 Geophysical Fluid Dynamics, FOS: Earth and related environmental sciences, FOS: Earth and related environmental sciences, 20303 Fluid Physics, FOS: Physical sciences, FOS: Physical sciences, 91501 Computational Fluid Dynamics, 90101 Aerodynamics (excl. Hypersonic Aerodynamics)},
+  keywords  = {Mechanical Engineering, FOS: Mechanical engineering, 91504 Fluidisation and Fluid Mechanics, FOS: Other engineering and technologies, 40403 Geophysical Fluid Dynamics, FOS: Earth and related environmental sciences, 20303 Fluid Physics, FOS: Physical sciences, 91501 Computational Fluid Dynamics, 90101 Aerodynamics (excl. Hypersonic Aerodynamics)},
   publisher = {figshare},
 }
 
@@ -72,11 +73,8 @@
   publisher = {Ubiquity Press, Ltd.},
 }
 
-
-
-
 @InProceedings{Ratz2022a,
-  author    = {Ratz, M., Sachs, S., K\"{o}nig, J., Mendez, M. A., Cierpka, C.},
+  author    = {M. Ratz and S. Sachs and J. K\"{o}nig and M. A. Mendez and C. Cierpka.},
   booktitle = {International Symposium on Applications of Laser and Imaging Techniques to Fluid Mechanics},
   title     = {{Radial basis function regression of Lagrangian three-dimensional particle tracking data}},
   year      = {2022},
@@ -144,3 +142,4 @@
   publisher = {Springer Science and Business Media {LLC}},
 }
 
+@Comment{jabref-meta: databaseType:bibtex;}
