### 2024-01-09 #5749 @mendezVKI
Hello Philippe,

The article looks good. The current release in PyPi is the final one. The folder was uploaded on Zenodo here:

https://zenodo.org/records/10473329
