### 2024-01-09 #5749 @philipcardiff
> Hello Philippe,
>
> The article looks good. We just uploaded a minor update to correct the display of the German "ö" in a bib entry. The current release in PyPi is the final one; we just tested and works. The folder was uploaded on Zenodo here:
>
> https://zenodo.org/records/10473329

Thanks @mendezVKI. Could you add a tag to the main SPICY_VKI GitHub repository? i.e. add a "1.0.5" tag for the latest code. In this way, users can be sure they are using the exact repository in this article. Thanks.
