### 2024-01-09 #5749 @philipcardiff
FYI, I proposed some minor English edits to the paper at https://github.com/mendezVKI/SPICY_VKI/pull/10
