### 2024-01-15 #5749 @kyleniemeyer
@mendezVKI the issue looks to be in your `paper.bib` file, specifically the `Heyman2019` reference. Please change the `journal` field to be just `Computers \& Geosciences`.
