+++ paper/paper.bib
@@ -23,11 +23,9 @@
 }
 
 
-
-
 @Article{Heyman2019,
   author    = {Joris Heyman},
-  journal   = {Computers {\&}amp$\mathsemicolon$ Geosciences},
+  journal   = {Computers \& Geosciences},
   title     = {{TracTrac}: A fast multi-object tracking algorithm for motion estimation},
   year      = {2019},
   month     = {jul},
