### 2024-01-15 255759f
Updates JOSS paper
fixes some reference commands and other minor edits
